+++ src/horizontal_missile.c
@@ -4,7 +4,7 @@
 #include "strategy.h"
 #include "display_macros.h"
 
-extern unsigned int loop;
+extern unsigned short loop;
 extern unsigned char level;
 
 extern Character leftEnemyMissile;
